Add simple photo browser
File system folders on the left side. View in the center, split up into
a foto roll of the folder at the bottom and display of the selection
above.

Current limitations:

Only contents of the selected folder are shown, non-recursive.
Shows thumbnails only from EXIF data in the foto roll.

Viewer is static, showing the picture scaled to fit.
Does not support videos or SVG.

All operations are blocking, no threading done yet. This does not lead
to long UI blocking, but sometimes doesn't feel fast enough.

State is not restored on startup.