Show (macOS style) file tags
These are saved as extended attributes of the files. The attribute
contains a (binary) plist with a list of strings. For "color" tags, the
string contains a line break, with the name before the line break and
the "color number" after the line break (I guess for translations?).

The implementation uses a 3rdparty plist library, and through that is
independent of macOS APIs, so it should in principle also work on Linux.